requirements.txt
Flask==3.0.0
openai==0.28.0
pandas==2.1.2
pydub==0.25.1
Requests==2.31.0
Werkzeug==3.0.1
azure-identity==1.17.1
azure-cognitiveservices-speech==1.38.0
python-dotenv==1.0.1